+++ src/app/earthquake/earthquake-resolver.service.ts
