+++ src/app/earthquake/earthquake.component.html
