+++ src/app/earthquake/earthquake.component.scss
