+++ src/app/earthquake/earthquake.guard.ts
