+++ src/app/earthquake/earthquake.service.ts
