+++ src/app/earthquake/earthquake.spec.ts
