+++ src/app/earthquake/earthquakedata.ts
